+++ packages/core/test/test-util/example-destination-service-responses.ts
@@ -66,6 +66,38 @@ export const oauthSingleResponse: DestinationJson = {
   ]
 };
 
+export const oauthUserTokenExchangeMultipleResponse: DestinationConfiguration[] = [
+  {
+    Name: 'FINAL-DESTINATION',
+    Type: 'HTTP',
+    URL: 'https://my.system.com/',
+    Authentication: 'OAuth2UserTokenExchange' as AuthenticationType,
+    ProxyType: 'Internet',
+    audience: 'https://my.system.com',
+    clientKey: 'password',
+    scope: 'SOME_SCOPE',
+    tokenServiceUser: 'TOKEN_USER',
+    tokenServiceURL: 'https://my.system.com/sap/bc/sec/oauth2/token',
+    userIdSource: 'email',
+    tokenServicePassword: 'password'
+  }
+];
+
+export const oauthUserTokenExchangeSingleResponse: DestinationJson = {
+  owner: {
+    SubaccountId: 'a89ea924-d9c2-4eab-84fb-3ffcaadf5d24',
+    InstanceId: null
+  },
+  destinationConfiguration: oauthUserTokenExchangeMultipleResponse[0],
+  authTokens: [
+    {
+      type: 'Bearer',
+      value: 'token',
+      expires_in: '3600'
+    }
+  ]
+};
+
 export const onPremiseMultipleResponse: DestinationConfiguration[] = [
   {
     Name: 'OnPremise',